chore(AutoTestkit): remove code examples
as decided, code examples should not be part of AutoTestkit and should
be held separately
