+++ packages/wix-storybook-utils/src/AutoTestkit/index.test.tsx
@@ -69,23 +69,5 @@ describe('AutoTestkit', () => {
     it('has driver documentation tables', () => {
       expect(driver.get.driverAt(0).get.name()).toBe(component.drivers[0].file);
     });
-
-    describe('code examples', () => {
-      it('renders enzyme example', () => {
-        expect(driver.get.codeExampleAt(0).get.type()).toEqual('enzyme');
-      });
-
-      it('renders protractor example', () => {
-        expect(driver.get.codeExampleAt(1).get.type()).toEqual('protractor');
-      });
-
-      it('renders puppeteer example', () => {
-        expect(driver.get.codeExampleAt(2).get.type()).toEqual('puppeteer');
-      });
-
-      it('defaults to enzyme example for unknown type', () => {
-        expect(driver.get.codeExampleAt(3).get.type()).toEqual('enzyme');
-      });
-    });
   });
 });
