+++ packages/wix-storybook-utils/src/Sections/views/testkit.tsx
@@ -2,11 +2,9 @@ import * as React from 'react';
 
 import { TestkitSection } from '../../typings/story-section';
 import { StoryConfig } from '../../typings/story-config';
-import { AutoTestkit } from '../../AutoTestkit/auto-testkit';
+import { AutoTestkit } from '../../AutoTestkit';
 
 export const testkit = (
   section: TestkitSection,
   storyConfig: StoryConfig,
-): React.ReactNode => (
-  <AutoTestkit unidriver={section.unidriver} component={storyConfig.metadata} />
-);
+): React.ReactNode => <AutoTestkit component={storyConfig.metadata} />;
